create scenarios file
