+++ rust/otel-arrow-rust/src/lib.rs
@@ -21,7 +21,6 @@ mod otlp;
 mod schema;
 #[cfg(test)]
 mod test_util;
-#[cfg(test)]
 mod validation;
 
 pub mod pdata;
